refactor(paths): move read_css_dir to paths
