+++ crates/igneous-md/src/paths.rs
@@ -1,15 +1,14 @@
-//! External paths for the application
+//! External paths and related functionality
 //!
-//! The [Paths] struct is the only state managed by the [rocket] server, since it is needed to init
+//! The [Paths] struct is State managed by the [rocket] server, since it is needed to create
 //! new [crate::client::Client]s.
 use home::home_dir;
+use itertools::Itertools;
 use std::{
-    io,
+    fs, io,
     path::{Path, PathBuf},
     sync::OnceLock,
 };
-
-use crate::config::read_css_dir;
 
 /// Returns the default config dir for the application
 pub fn default_config_dir() -> &'static PathBuf {
@@ -25,6 +24,40 @@ pub fn default_config_dir() -> &'static PathBuf {
 pub fn default_css_dir() -> &'static PathBuf {
     static CSS_PATH: OnceLock<PathBuf> = OnceLock::new();
     CSS_PATH.get_or_init(|| default_config_dir().join("css"))
+}
+
+/// Will attempt to read the given `css_dir` and organize the output
+///
+/// This function will:
+///
+/// 1. Only include `.css` files
+///
+/// 2. Return only names prefixed with `/css`
+///
+/// 3. Sort them by their name
+pub fn read_css_dir(css_dir: &Path) -> io::Result<Vec<PathBuf>> {
+    Ok(fs::read_dir(css_dir)?
+        .filter_map(|possible_entry| {
+            let path = possible_entry.ok()?.path();
+
+            if path.is_file() && path.extension().is_some_and(|s| s == "css") {
+                return Some(
+                    PathBuf::from("/css").join(
+                        path.strip_prefix(css_dir)
+                            .expect("We read the files from the css_dir."),
+                    ),
+                );
+            }
+
+            None
+        })
+        .sorted_by_key(|p| {
+            PathBuf::from(
+                p.file_name()
+                    .expect("We checked that all entries are files."),
+            )
+        })
+        .collect())
 }
 
 /// Paths used by the application
